change to normal double

+++ PIP200644.h
@@ -4,12 +4,12 @@ namespace N200644 {
 
 class Point {
 private:
-    long double _x, _y;
+    double _x, _y;
 public:
     Point() { _x = 0; _y = 0; }
     Point(double x, double y) { _x = x; _y = y; }
-    long double x() const {return _x;}
-    long double y() const {return _y;}
+    double x() const {return _x;}
+    double y() const {return _y;}
 };
 
 class Polygon {
@@ -40,7 +40,7 @@ Polygon::Polygon(InputIterator first, InputIterator last){
 
 /* Check if a point is on the left (>0) or right (<0) side, based on the cross product.
  * The result is exactly 0 when q is part of the line spanned by head and tail */  
-inline long double checkSide( const Point q, const Point head, const Point tail){
+inline double checkSide( const Point q, const Point head, const Point tail){
     return ( (head.x()-tail.x())*(q.y()-tail.y()) - (q.x()-tail.x())*(head.y()-tail.y()) );
 }
